Simplify htmlSafe declaration
This was confusing to read and was deep nested in a way that wasn't
required. Pulling this up into a single nested destructor and one line
made it easy to see what the intent of the statement was for.

Perhaps if you wanted to destruct more Ember.String prototype functions
you might consider such a pattern but in this case it was not needed and
added cognitive complexity.